Project file may be outdated with regard to name
When one copies a dtsx package in the solution explorer. The name is not updated right away in the updated package. One has to open the package and change something or rename it. Otherwise the project contains two packages with the same Name attribute in dtproj file. This made ssis-build.exe crash.